fix(yttrex): guardoni uses v2 endpoints and experiment registration works
