+++ platforms/yttrex/shared/src/endpoints/v2/index.ts
@@ -1,9 +1,11 @@
 import Metadata from './metadata.endpoints';
 import Public from './public.endpoints';
+import Experiments from './experiments.endpoints';
 
 const endpoints = {
   Public,
   Metadata,
+  Experiments,
 };
 
 export default endpoints;
