+++ platforms/yttrex/shared/src/endpoints/v3/public.endpoints.ts
@@ -1,24 +1,10 @@
 import {
-  CreateExperimentResponse, GetPublicDirectivesOutput
-} from '@shared/models/Experiment';
-import { HandshakeBody } from '@shared/models/HandshakeBody';
-import {
   GetRecommendationsParams,
   GetRecommendationsQuery,
   RecommendationList
 } from '@shared/models/Recommendation';
-import { Step } from '@shared/models/Step';
 import * as t from 'io-ts';
 import { Endpoint } from 'ts-endpoint';
-
-const Handshake = Endpoint({
-  Method: 'POST',
-  getPath: () => `/v3/handshake`,
-  Input: {
-    Body: HandshakeBody,
-  },
-  Output: t.any,
-});
 
 const VideoRecommendations = Endpoint({
   Method: 'GET',
@@ -44,40 +30,7 @@ const GetRecommendations = Endpoint({
   Output: RecommendationList,
 });
 
-const PostDirective = Endpoint({
-  Method: 'POST',
-  getPath: () => `/v3/directives`,
-  Input: {
-    Headers: t.type({
-      'Content-Type': t.string,
-    }),
-    Body: t.array(Step),
-  },
-  Output: CreateExperimentResponse,
-});
-
-const GetDirective = Endpoint({
-  Method: 'GET',
-  getPath: ({ experimentId }) => `/v3/directives/${experimentId}`,
-  Input: {
-    Params: t.type({
-      experimentId: t.string,
-    }),
-  },
-  Output: t.array(Step),
-});
-
-const GetPublicDirectives = Endpoint({
-  Method: 'GET',
-  getPath: () => `/v3/directives/public`,
-  Output: GetPublicDirectivesOutput,
-});
-
 export const endpoints = {
-  Handshake,
   GetRecommendations,
   VideoRecommendations,
-  GetDirective,
-  PostDirective,
-  GetPublicDirectives,
 };
